+++ dbt/include/fabric/macros/dbt_package_support/dbt_utils/sql/get_tables_by_pattern_sql.sql
@@ -1,16 +1,20 @@
-{#- Override: queries sys.tables joined with sys.schemas instead of information_schema.tables.
-    T-SQL sys views provide reliable pattern matching via LIKE on Fabric Data Warehouse. -#}
 {% macro fabric__get_tables_by_pattern_sql(schema_pattern, table_pattern, exclude='', database=target.database) %}
 
-        select
-                s.name as table_schema,
-                t.name as table_name,
-                'table' as table_type
-        from sys.tables t
-        inner join sys.schemas s
-        on s.schema_id = t.schema_id
-        where s.name like '{{ schema_pattern }}'
-        and t.name like '{{ table_pattern }}'
-        and t.name not like '{{ exclude }}'
+    select
+        s.name as table_schema,
+        o.name as table_name,
+        case o.type
+            when 'U' then 'table'
+            when 'V' then 'view'
+        end as table_type
+    from sys.objects as o
+    inner join sys.schemas as s
+        on s.schema_id = o.schema_id
+    where o.type in ('U', 'V')
+      and s.name like '{{ schema_pattern }}'
+      and o.name like '{{ table_pattern }}'
+      {% if exclude %}
+      and o.name not like '{{ exclude }}'
+      {% endif %}
 
 {% endmacro %}